Freetype library update Engine - FileWatcher

+++ Editor/EditorLayers/ImGuiLayer.cpp
@@ -18,8 +18,7 @@ void ImGuiLayer::OnAttach() {
     ImGui::CreateContext();
     ImGuiIO& io = ImGui::GetIO(); (void)io;
     io.ConfigFlags |= ImGuiConfigFlags_NavEnableKeyboard;     // Enable Keyboard Controls
-    //io.DisplayFramebufferScale = ImVec2(1.25f, 1.25f);
-    io.DisplaySize = ImVec2(win->GetWindowWidth(), win->GetWindowWidth());
+    io.DisplaySize = ImVec2((float)win->GetWindowWidth(), (float)win->GetWindowWidth());
     //io.DisplayFramebufferScale = ImVec2(1.25f, 1.25f);
 
     // Setup Dear ImGui style
